bug sur delete post

+++ backend/controllers/post-controllers.js
@@ -1,6 +1,7 @@
 const Post = require('../models/post-models');
 const User = require('../models/user-models');
-const fs = require('fs')
+const fs = require('fs');
+const { post } = require('../routes/post-routes');
 
 exports.createPost = (req, res, next) => {
     if (req.file) {
@@ -58,15 +59,23 @@ exports.modifyPost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
     Post.findOne({ where: { id: req.params.id }})
         .then(post => {
-            const filename = post.image.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
+            if (post.image != null) {
+                const filename = post.image.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Post.destroy({ where: { id: req.params.id } })
+
+                    .then(() => res.status(200).json({message : 'Post supprimé !'}))
+                    .catch( error => res.status(400).json({error}));
+                })
+            
+        
+            } else {
                 Post.destroy({ where: { id: req.params.id } })
 
                 .then(() => res.status(200).json({message : 'Post supprimé !'}))
                 .catch( error => res.status(400).json({error}));
-            })
-        }
-        )
+            }
+        })
 }
 
 exports.getAllPosts = (req, res, next) => {
